+++ src/module/allocation/allocation.controller.js
@@ -22,22 +22,7 @@ const onLoadCarrierAllocation = (req, res) => {
     allocationService.onLoadCarrierAllocation(req, res).then((response) => res.send(response))
 }
 
-const quotationCompanyDetails = (req, res) => {
-    logger.info(`Quoation Company Details Page API called. ${JSON.stringify(req.body)}`)
-    // allocationService.quotationCompanyDetails(req, res).then((response) => res.send(response))
-    allocationService.quotationCompanyDetails(req, res)
-}
 
-const insertQuotationCompanyDetails = (req, res) => {
-    logger.info(`Insert Quoation Company Details Page API called. ${JSON.stringify(req.body)}`)
-    // allocationService.quotationCompanyDetails(req, res).then((response) => res.send(response))
-    allocationService.insertQuotationCompanyDetails(req, res)
-}
-const updateQuotationCompanyDetails = (req, res) => {
-    logger.info(`Update Quoation Company Details Page API called. ${JSON.stringify(req.body)}`)
-    // allocationService.quotationCompanyDetails(req, res).then((response) => res.send(response))
-    allocationService.updateQuotationCompanyDetails(req, res)
-}
 const updateTargetValues = (req, res) => {
     logger.info(`Update target values of Carrier allocation ${JSON.stringify(req.body)}`)
     allocationService.updateTargetValues(req, res).then((response) => res.send(response))
@@ -54,7 +39,4 @@ module.exports = {
     onLoadCarrierAllocation,
     updateTargetValues,
     downloadExcel,
-    quotationCompanyDetails,
-    insertQuotationCompanyDetails,
-    updateQuotationCompanyDetails
 };
