+++ src/module/dashboard/dashboard.route.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router({ mergeParams: true });
 
 const dashboardController = require('./dashboard.controller');
-const { dashboardBasePath, QuotationBasePath, QuotationCreateBasePath, QuotationUpdateBasePath, QuotationBasePathById ,OnLoadQoutation} = require('../../utils/urlConstant');
+const { dashboardBasePath, QuotationBasePath, QuotationCreateBasePath, QuotationUpdateBasePath, QuotationBasePathById, OnLoadQoutation, quotationCompanyDetails, insertQuotationCompanyDetails, updateQuotationCompanyDetails } = require('../../utils/urlConstant');
 
 router.route(dashboardBasePath).post(dashboardController.getDataFromExcelSheet);
 
@@ -13,5 +13,9 @@ router.route(QuotationCreateBasePath).post(dashboardController.createQuotation);
 router.route(QuotationUpdateBasePath).post(dashboardController.updateQuotation);
 router.route(QuotationUpdateBasePath).post(dashboardController.updateQuotation);
 router.route(OnLoadQoutation).get(dashboardController.onLoadQuotations);
+
+router.route(quotationCompanyDetails).get(dashboardController.quotationCompanyDetails);
+router.route(insertQuotationCompanyDetails).post(dashboardController.insertQuotationCompanyDetails);
+router.route(updateQuotationCompanyDetails).post(dashboardController.updateQuotationCompanyDetails);
 
 module.exports = router;
