+++ src/utils/urlConstant.js
@@ -30,9 +30,6 @@ module.exports.onLoadCarrierAllocation = "/carrier/onload";
 
 module.exports.downloadExcelPath = "/carrier/download";
 
-module.exports.quotationCompanyDetails = "/quotation-details";
-module.exports.insertQuotationCompanyDetails = "/insert-quotation-details";
-module.exports.updateQuotationCompanyDetails = "/update-quotation-details";
 //Dashboard Endpoint
 module.exports.dashboardBasePath = "/dashboard"
 module.exports.QuotationBasePath = "/quotation/fetch"
@@ -40,6 +37,9 @@ module.exports.QuotationBasePathById = "/quotation/fetchById"
 module.exports.QuotationCreateBasePath = "/quotation/create"
 module.exports.QuotationUpdateBasePath = "/quotation/update"
 module.exports.OnLoadQoutation = "/quotation/on-load"
+module.exports.quotationCompanyDetails = "/quotation-details";
+module.exports.insertQuotationCompanyDetails = "/insert-quotation-details";
+module.exports.updateQuotationCompanyDetails = "/update-quotation-details";
 
 // Customer module endpoint
 module.exports.customerbaseUrl = "/customer/v1"
